Allow for cases where the default frequency is too high
I spent a couple of hours trying to figure out if I'd lost my mind until
I came across this article which suggested lowering the frequency to get
past "OSError: [Errno 5] EIO".

This makes frequency settable in the test script and adds a note to the
README as well. FWIW I'm using the Keyestudio LCM1602.